added components.js file
