+++ src/config.js
@@ -10,8 +10,18 @@ module.exports = {
         localAreasMap.set("Columbus", "ChIJcd6QucGJOIgRM7Wxz_hmMuQ");
         localAreasMap.set("Dublin", "ChIJH6FQ1MTsOIgRKJBoFWgXwgA");
         localAreasMap.set("Hilliard", "ChIJMxtWksaWOIgRnlXah9jo_aE");
-        localAreasMap.set("Westerville", "ChIJVyNMY2X1OIgRQT9dsFQwoUY");    
+        localAreasMap.set("Westerville", "ChIJVyNMY2X1OIgRQT9dsFQwoUY");
         return localAreasMap;
+    },
+
+    ZipCodes() {
+        let zipCodeMap = new Map();
+        zipCodeMap.set("43201", "ChIJ9Rz24rWOOIgR3EEuL2Ge4oo");
+        zipCodeMap.set("43206", "ChIJw6wuB5iIOIgRus0jBTvKLHo");
+        zipCodeMap.set("43212", "ChIJOW2nsVCOOIgR0IzLWzB30l4");
+        zipCodeMap.set("43215", "ChIJmRPPvRCGOIgRARpJZtkXzq0");
+        zipCodeMap.set("43224", "ChIJj3ATYruLOIgRMmVSsRFsuQ0");
+        return zipCodeMap;
     }
 
 }
